Remove temporary file
