+++ src/autoresearch/distributed/__init__.py
@@ -1,0 +1,25 @@
+"""Distributed execution utilities."""
+
+from .broker import InMemoryBroker, RedisBroker, BrokerType, get_message_broker
+from .coordinator import (
+    StorageCoordinator,
+    ResultAggregator,
+    start_storage_coordinator,
+    start_result_aggregator,
+    publish_claim,
+)
+from .executors import RayExecutor, ProcessExecutor
+
+__all__ = [
+    "InMemoryBroker",
+    "RedisBroker",
+    "BrokerType",
+    "get_message_broker",
+    "StorageCoordinator",
+    "ResultAggregator",
+    "start_storage_coordinator",
+    "start_result_aggregator",
+    "publish_claim",
+    "RayExecutor",
+    "ProcessExecutor",
+]
