+++ src/autoresearch/search/__init__.py
@@ -1,0 +1,13 @@
+"""Search subpackage."""
+
+from .core import Search
+from .context import SearchContext
+from .http import get_http_session, set_http_session, close_http_session
+
+__all__ = [
+    "Search",
+    "SearchContext",
+    "get_http_session",
+    "set_http_session",
+    "close_http_session",
+]
