+++ src/autoresearch/search/core.py
@@ -77,332 +77,16 @@ except ImportError:
     BERTOPIC_AVAILABLE = False
 
 import atexit
-from .config import get_config
 from .errors import ConfigError, SearchError
 from .logging_utils import get_logger
 from .cache import get_cached_results, cache_results
+from .config import get_config
+log = get_logger(__name__)
 from .storage import StorageManager
 
-_http_session: Optional[requests.Session] = None
-_http_lock = threading.Lock()
-_atexit_registered = False
+from .http import get_http_session, close_http_session, set_http_session
+from .context import SearchContext
 
-
-def set_http_session(session: requests.Session) -> None:
-    """Inject an existing HTTP session (for distributed workers)."""
-    global _http_session, _atexit_registered
-    with _http_lock:
-        _http_session = session
-        if not _atexit_registered:
-            atexit.register(close_http_session)
-            _atexit_registered = True
-
-
-def get_http_session() -> requests.Session:
-    """Return a pooled HTTP session."""
-    global _http_session, _atexit_registered
-    with _http_lock:
-        if _http_session is None:
-            cfg = get_config()
-            size = getattr(cfg.search, "http_pool_size", 10)
-            session = requests.Session()
-            adapter = requests.adapters.HTTPAdapter(
-                pool_connections=size, pool_maxsize=size
-            )
-            session.mount("http://", adapter)
-            session.mount("https://", adapter)
-            _http_session = session
-            if not _atexit_registered:
-                atexit.register(close_http_session)
-                _atexit_registered = True
-        return _http_session
-
-
-def close_http_session() -> None:
-    """Close the pooled HTTP session."""
-    global _http_session, _atexit_registered
-    with _http_lock:
-        if _http_session is not None:
-            _http_session.close()
-            _http_session = None
-            _atexit_registered = False
-
-
-log = get_logger(__name__)
-
-
-class SearchContext:
-    """Manages context for context-aware search.
-
-    This class maintains the context for search queries, including:
-    - Search history
-    - Extracted entities
-    - Topic models
-    - Query expansions
-
-    It provides methods for:
-    - Adding queries to the search history
-    - Extracting entities from text
-    - Building topic models
-    - Expanding queries based on context
-    """
-
-    _instance = None
-    _lock = threading.Lock()
-
-    @classmethod
-    def get_instance(cls) -> "SearchContext":
-        """Get or create the singleton instance of SearchContext.
-
-        Returns:
-            SearchContext: The singleton instance
-        """
-        with cls._lock:
-            if cls._instance is None:
-                cls._instance = SearchContext()
-            return cls._instance
-
-    def __init__(self):
-        """Initialize the search context."""
-        self.search_history = []
-        self.entities = defaultdict(int)
-        self.topic_model = None
-        self.dictionary = None
-        self.nlp = None
-        self._initialize_nlp()
-
-    def _initialize_nlp(self) -> None:
-        """Initialize the NLP model if available."""
-        if not SPACY_AVAILABLE:
-            return
-
-        try:
-            self.nlp = spacy.load("en_core_web_sm")
-            log.info("Initialized spaCy NLP model")
-        except OSError:
-            if (
-                os.getenv("AUTORESEARCH_AUTO_DOWNLOAD_SPACY_MODEL", "").lower()
-                == "true"
-            ):
-                try:
-                    spacy.cli.download("en_core_web_sm")
-                    self.nlp = spacy.load("en_core_web_sm")
-                    log.info("Downloaded spaCy model")
-                except Exception as e:  # pragma: no cover - unexpected
-                    log.warning(f"Failed to download spaCy model: {e}")
-            else:
-                log.warning(
-                    "spaCy model 'en_core_web_sm' not found. Set "
-                    "AUTORESEARCH_AUTO_DOWNLOAD_SPACY_MODEL=true to download automatically."
-                )
-        except Exception as e:  # pragma: no cover - unexpected
-            log.warning(f"Failed to initialize spaCy NLP model: {e}")
-
-    def add_to_history(self, query: str, results: List[Dict[str, str]]):
-        """Add a query and its results to the search history.
-
-        Args:
-            query: The search query
-            results: The search results
-        """
-        cfg = get_config()
-        max_history = cfg.search.context_aware.max_history_items
-
-        # Add to history
-        self.search_history.append(
-            {"query": query, "results": results, "timestamp": time.time()}
-        )
-
-        # Trim history if needed
-        if len(self.search_history) > max_history:
-            self.search_history = self.search_history[-max_history:]
-
-        # Extract entities from query and results
-        self._extract_entities(query)
-        for result in results:
-            self._extract_entities(result.get("title", ""))
-            self._extract_entities(result.get("snippet", ""))
-
-    def _extract_entities(self, text: str):
-        """Extract entities from text and update the entity counts.
-
-        Args:
-            text: The text to extract entities from
-        """
-        if not SPACY_AVAILABLE or self.nlp is None:
-            return
-
-        try:
-            doc = self.nlp(text)
-            for ent in doc.ents:
-                if ent.label_ in [
-                    "PERSON",
-                    "ORG",
-                    "GPE",
-                    "LOC",
-                    "PRODUCT",
-                    "EVENT",
-                    "WORK_OF_ART",
-                    "LAW",
-                ]:
-                    self.entities[ent.text.lower()] += 1
-        except Exception as e:
-            log.warning(f"Entity extraction failed: {e}")
-
-    def build_topic_model(self):
-        """Build a topic model from the search history."""
-        if (
-            not BERTOPIC_AVAILABLE
-            or not self.search_history
-            or not SENTENCE_TRANSFORMERS_AVAILABLE
-        ):
-            return
-
-        try:
-            # Collect all text from search history
-            documents = []
-            for item in self.search_history:
-                query = item["query"]
-                documents.append(query)
-
-                for result in item["results"]:
-                    title = result.get("title", "")
-                    snippet = result.get("snippet", "")
-                    if title:
-                        documents.append(title)
-                    if snippet:
-                        documents.append(snippet)
-
-            # Check if we have enough documents
-            if len(documents) < 2:
-                log.warning("Not enough documents to build a topic model")
-                return
-
-            # Get sentence transformer model
-            sentence_transformer = Search.get_sentence_transformer()
-            if sentence_transformer is None:
-                log.warning("Sentence transformer model not available")
-                return
-
-            # Build BERTopic model
-            cfg = get_config()
-            num_topics = cfg.search.context_aware.num_topics
-
-            # Initialize BERTopic with the sentence transformer model
-            self.topic_model = BERTopic(
-                embedding_model=sentence_transformer, nr_topics=num_topics
-            )
-
-            # Fit the model on the documents
-            self.topic_model.fit(documents)
-
-            # Store the documents for later use
-            self.documents = documents
-
-            log.info(f"Built topic model with {num_topics} topics")
-        except Exception as e:
-            log.warning(f"Topic modeling failed: {e}")
-
-    def expand_query(self, query: str) -> str:
-        """Expand a query based on context.
-
-        Args:
-            query: The original query
-
-        Returns:
-            str: The expanded query
-        """
-        cfg = get_config()
-        context_cfg = cfg.search.context_aware
-
-        if not context_cfg.enabled or not context_cfg.use_query_expansion:
-            return query
-
-        expanded_terms = []
-
-        # Add terms from entities
-        if (
-            context_cfg.use_entity_recognition
-            and SPACY_AVAILABLE
-            and self.nlp is not None
-        ):
-            # Extract entities from the query
-            try:
-                doc = self.nlp(query)
-                query_entities = [ent.text.lower() for ent in doc.ents]
-
-                # Find related entities from history
-                for entity, count in sorted(
-                    self.entities.items(), key=lambda x: x[1], reverse=True
-                ):
-                    if entity not in query_entities and count > 1:
-                        expanded_terms.append(entity)
-                        if len(expanded_terms) >= 3:  # Limit to top 3 entities
-                            break
-            except Exception as e:
-                log.warning(f"Entity-based query expansion failed: {e}")
-
-        # Add terms from topic model
-        if (
-            context_cfg.use_topic_modeling
-            and BERTOPIC_AVAILABLE
-            and self.topic_model is not None
-            and hasattr(self, "documents")
-        ):
-            try:
-                # Get sentence transformer model
-                sentence_transformer = Search.get_sentence_transformer()
-                if sentence_transformer is None:
-                    log.warning(
-                        "Sentence transformer model not available for query expansion"
-                    )
-                else:
-                    # Get the topic for the query
-                    topic, _ = self.topic_model.transform([query])
-                    topic = topic[0]  # Get the first (and only) topic
-
-                    # If the topic is valid (not -1, which means no topic)
-                    if topic != -1:
-                        # Get the top terms for the topic
-                        topic_info = self.topic_model.get_topic(topic)
-                        for term, _ in topic_info[:3]:  # Get top 3 terms
-                            if (
-                                term.lower() not in query.lower()
-                                and term not in expanded_terms
-                            ):
-                                expanded_terms.append(term)
-            except Exception as e:
-                log.warning(f"Topic-based query expansion failed: {e}")
-
-        # Add terms from search history
-        if context_cfg.use_search_history and self.search_history:
-            try:
-                # Get the most recent search query
-                recent_query = self.search_history[-1]["query"]
-
-                # Add terms from recent query that aren't in the current query
-                recent_terms = Search.preprocess_text(recent_query)
-                query_terms = Search.preprocess_text(query)
-
-                for term in recent_terms:
-                    if term not in query_terms and term not in expanded_terms:
-                        expanded_terms.append(term)
-                        if len(expanded_terms) >= 2:  # Limit to top 2 terms
-                            break
-            except Exception as e:
-                log.warning(f"History-based query expansion failed: {e}")
-
-        # Combine original query with expanded terms
-        if expanded_terms:
-            expansion_factor = context_cfg.expansion_factor
-            num_terms = max(1, int(len(expanded_terms) * expansion_factor))
-            selected_terms = expanded_terms[:num_terms]
-
-            expanded_query = f"{query} {' '.join(selected_terms)}"
-            log.info(f"Expanded query: '{query}' -> '{expanded_query}'")
-            return expanded_query
-
-        return query
 
 
 class Search:
